+++ src/config.ts
@@ -14,6 +14,9 @@ export async function get(
   key: "playground.autoRun"
 ): Promise<"onEdit" | "onSave" | "never">;
 export async function get(key: "playground.autoSave"): Promise<boolean>;
+export async function get(
+  key: "playground.layout"
+): Promise<"grid" | "splitLeft" | "splitRight" | "splitTop">;
 export async function get(key: "playground.includeMarkup"): Promise<boolean>;
 export async function get(
   key: "playground.includeStylesheet"