+++ src/interfaces/IPlaygroundJSON.ts
@@ -1,4 +1,6 @@
 export interface IPlaygroundJSON {
-  scripts: string[];
-  styles: string[];
+  scripts?: string[];
+  styles?: string[];
+  layout?: string;
+  showConsole?: boolean;
 }